Restructured src
So while testing the installation of the package using pip I cam to the conclusion that using our old file structure:

─── src
    ├── pychart
    |   ├── __init__.py
    |   ├── __main__.py
    |   └── runner.py
    ├── token_type
    |   ├── __init__.py
    |   ├── token.py
    |   └── token_type_enum.py
    ├── __init__.py
    ├── expressions.py
    ├── pyparser.py
    └── scanner.py

Is not possible because when install the package through pip
from `src` it installs all the modules in that directory, to
acctually make it work with we would have to install the
whole src package, which would mean that you would have to
write `import src.*` to import anything in our package
(another way to install it would be to install
`expressions.py, ...` as modules as well, but then if
someone installed our package they would also get access to
those modules anyway. This seems unacceptable because this
would remove any point of putting this submodules inside the
src module (as they would be accessible anyway, which is one
of the reasons for this file structure). According to my
understanding of python I don't think it's possible to
ensure this inecassibility, because installing python
packages litterally just downloads a compressed module and
uncompresses it.

As such I propose the following file structure:

─── src
    └── pychart
        ├── token_type
        |   ├── __init__.py
        |   ├── token.py
        |   └── token_type_enum.py
        ├── __init__.py
        ├── __main__.py
        |── runner.py
        ├── expressions.py
        ├── pyparser.py
        └── scanner.py

I have already updated all the imports.