Added SEPARATOR TokenType
Previously we discussed about ignoring whitspaces but I don't think that should be handeled by the scanner, because
we loose vital information which can be used for error messages, by introducing a SEPARATOR token we can turn the
tokens into the original source code if an error occurs.

Instead the SEPARATOR token itself should be ignored by the parsers.

I also added the scanning of the SEPARATOR token