Updated run_prompt to handle multiline statements
Previously multiline statements were not possible because the scanner automatically appends the EOF token

Originally I thought fixing this might involve complicate syntax checking like makin sure that all '(' are
closed etc.

I think this solution is the simplest to implement, that is every statement is ended with a empty input
(newline).